#1 Rename MySQL dogu to MariaDB
The existing MySQL code was not used as regular dogu - it was never officially
released. After years of stagnation the need for a MySQL dogu awakened. Because
MariaDB is the official open source branch of MySQL, this commit reflects the
Open Source nature of this repository and renames this repository and all its
internal representations (with some exceptions which will be discussed below).

This commit renames all binaries and volumes which include the term MySQL (or
lowercase denominations). There are currently three exclusions which must
continue to use the mysql-terminology for technical reasons:
- `/var/run/mysqld` - the directory which contains socket or lock files
- config entries like `[mysqld]`, and
- database-internal structures like `mysql.db` or `mysql.user`