#1 Add backup-consumer
This commit adds a backup-consumer script to the MariaDB dogu which receives
all dogu-consumer's service account credentials (username, database, password).
These credentials are generically gathered and supplied by cesapp while running
the backup. It is the backup-consumer's responsibility to fetch the credentials
and put it to appropriate use.

The backup-consumer script here simply uses the database and discards other
credentials simply because the respective mariadb dump application requires only
the database name being dumped.

In order to simplify testing (BATS has a problem with mocking binaries
containing dashes "-") this commit adds another symlink to a mariadb binary with
a more compatible underscore "_".