#1 Enable logging to docker log facilities
`mysqld_save` (or the respective MariaDB counterpart) is an old script file
that originaly took care of different config and safety measures. Anyhow, using
this script disables proper logging.

This commit adds a symlink to stdout so that the log lines get forwarded to
docker which captures all output. Furthermore this commit replaces calls to
`mariadbd_safe` (see above) to the proper binary in order to enable regular
logging.