#1 No longer store root password
Setting the MariaDB root account is an additional security measure. Generally
it should be impossible to connect from inside and outside the container to
MariaDB as root.

Because we should not use the root account we now actively forget the password.
All administration is possible via the mariadb account which initially is not
connectable from outside either (`docker exec` not counted in).