#1 Make RAM limit configurable and render config templates at start-up
This commit haves the dogu configure the RAM usage in order to apply MariaDB's
recommendation of using 80 % RAM. In order to achive this, the Docker container
creates a new directory `/etc/my.cnf.dogu.d` in which the templated config file
will be rendered. Using the existing directory `/etc/my.cnf.d` is not possible
because of file permissions.

This commit adds a detection for `memory_limit` misconfiguration which will
happen when with `cesapp edit-config` the memory_limit key is set but the
container will merely be restarted (instead of being re-created). Under this
condition docker will incorrectly allow several Petabytes of RAM, and the
MariaDB container will most likely fail during start-up. The added check does
not change the behavior but rathe adds an ERROR log line in order to guide the
administrator what to do next.